+++ modules/submitTransaction.ts
@@ -1,8 +1,8 @@
 import { ethers } from "ethers";
 import { ZuploContext, ZuploRequest, environment } from "@zuplo/runtime";
 import { createClient } from "@supabase/supabase-js";
-const verifyRequestLegitimityOnContract = require("../verification/verifyRequestLegitimityOnContract");
-const verifyIsAnAddress = require("../verification/verifyIsAnAddress");
+const verifyRequestLegitimityOnContract = require("./verification/verifyRequestLegitimityOnContract");
+const verifyIsAnAddress = require("./verification/verifyIsAnAddress");
 
 const { WALLET_PRIVATE_KEY, QUICKNODE_API_KEY, INFURA_API_KEY, SUPABASE_URL, SUPABASE_PASSWORD } = environment;
